// frontend/src/app/core/generated/circabc/model/pagedSearchNodes.ts
/**
 * CIRCABC REST API
 *
 * Contact: DIGIT-CIRCABC-SUPPORT@ec.europa.eu
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { SearchNode } from './searchNode';


export interface PagedSearchNodes { 
    data: Array<SearchNode>;
    /**
     * return the total amount of nodes, without paging so the UI can compute the number of pages 
     */
    total: number;
}